Tar file is added to YetiTar folder. It can be executed by double click.
In order to plot the graph, I tried to write the failing values to a
file. Plan is that plot button will plot the values on the graph after
reading from the generated text file.

* if the tar file is not in the same folder which contain the testing
.class files and the DSSR2 generated .java files then compilation gives
an error. So to keep things smooth for now keep .jar, .java and the
.class file to test in the same folder.

Work still remaining is to get the specific values for which the program
is failing and save it in some readable format for plot button.